+++ storage_utils/tests/test_utils.py
@@ -32,6 +32,9 @@ class TestIsSwiftPath(unittest.TestCase):
 
 
 class TestIsFilesystemPath(unittest.TestCase):
+    def test_is_posix_path(self):
+        self.assertEquals(storage_utils.is_posix_path, storage_utils.is_filesystem_path)
+
     def test_true(self):
         self.assertTrue(storage_utils.is_filesystem_path('my/posix/path'))